requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
click==8.2.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
executing==2.2.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
pyzmq==27.0.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.2
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
wcwidth==0.2.13
